requirements.txt
streamlit
groq
PyPDF2
langchain-text-splitters
sentence-transformers
scikit-learn
numpy
Pillow
pdf2image
pytesseract
python-dotenv